requirements.txt
fastapi==0.110.0
pydantic==2.6.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.29
uvicorn[standard]==0.29.0
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-jose[cryptografy]==3.3.0